Release 0.2.11.7
- Added command line: initgateway.  This inits the gateway so
MyAPIGateway initializes and scripts run on dedicated server as well as
clients.  Please ensure scripts are right before trying this, otherwise
a script can crash the server.
- Added command line: restartoncrash.  This command line attempts to
restart the server if there is a crash (20 seconds after crash).  Please
note it's a bit of hack (uses .bat files).  Unfortunately there's no way
to cleanup the game properly to restart it without restarting the entire
process.
- Base class of a plugin now only required to implement IPlugin, and
isn't required to subclass PluginBase
- Added SandboxGameAssemblyWrapper.Instance.GameAction() to allow a
synchronous action to be executed in the game thread.  This allows you
to safely use MyAPIGateway.
- Added SandboxGameAssemblyWrapper.Instance.BeginGameAction() to allow
an asynchronous action to be executed in the game thread.  This allows
you to safely use MyAPIGateway and sets a callback to be called when the
action is executed.